refactor: modal 블러처리 수정
