+++ src/pages/MyPot/components/ChangeStatusModal/ChangeStatusModalWrapper/ChangeStatusModalWrapper.style.ts
@@ -1,0 +1,14 @@
+import { css } from "@emotion/react";
+
+export const changeStatusModalOverlayStyle = css`
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  z-index: 9999;
+  background-color: rgba(0, 0, 0, 0.4);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
